Begin test cleanup for polymorphic tests
